Add bulk edit toggle switch
